Use a regex to search files for mangled symbols
This greatly expands the usability of the crate,
by allowing users to automatically demangle all symbols in a file.
This was already a feature of c++filt, but wasn't in rustfilt.

In addition, rustfilt can now strip hashes from the demangled symbols,
and does so by default. This is usually what the user wants, as it makes
the symbols more readable, but it can be disabled with the -h flag.

Thanks to the speed of the regex crate, this is _very_ fast.
To demangle a 69MB file took 0.5s, which is 6.8x faster than c++filt.

Fixes #1