Simplify the demangling buffer management
The prior code assumed that the demangled output was never longer than
the input.  This isn't true, as the newly-added test case demonstrates.
It's simpler to just let `format!` build the string for us, and still
plenty fast.  Fixes #3.

Also, when writing the output, there's no need to convert the regex-
replaced `Cow` to an owned string.  We can just write the bytes as-is.